+++ autonomi-cli/src/commands.rs
@@ -9,7 +9,6 @@
 mod file;
 mod register;
 mod vault;
-use std::path::PathBuf;
 
 use clap::Subcommand;
 use color_eyre::Result;
@@ -56,7 +55,7 @@ pub enum FileCmd {
         /// The address of the file to download.
         addr: String,
         /// The destination file path.
-        dest_file: PathBuf,
+        dest_file: String,
     },
 
     /// List previous uploads